# openapi.json
{"openapi": "3.1.0", "info": {"title": "AppServerX", "version": "1.0.0"}, "paths": {"/bootstrap": {"get": {"tags": ["Common"], "summary": "Api Bootstrap", "description": "\u5ba2\u6237\u7aef\u521d\u59cb\u5316\u914d\u7f6e\u63a5\u53e3\u3002\n\n\u8fd4\u56de\u542f\u52a8\u53c2\u6570\u3001\u533a\u57df\u8bbe\u7f6e\u3001\u521d\u59cb\u6a21\u677f\u4e0e\u7f13\u5b58\u63a7\u5236\u4fe1\u606f\u3002", "operationId": "api_bootstrap", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/global-configuration": {"get": {"tags": ["Common"], "summary": "Api Global Configuration", "description": "\u83b7\u53d6\u5168\u5c40\u914d\u7f6e\u3002\n\n\u901a\u8fc7\u7b7e\u540d\u53c2\u6570\u6821\u9a8c\u540e\uff0c\u8fd4\u56de\u8fdc\u7a0b\u5168\u5c40\u914d\u7f6e\u4e2d\u5fc3\u914d\u7f6e\u7ed3\u679c\u3002", "operationId": "api_global_configuration", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/keepalive-render": {"get": {"tags": ["Common"], "summary": "Api Keepalive Render", "description": "\u9632 Render \u4f11\u7720\u63a5\u53e3\u3002\n\n\u901a\u8fc7\u6267\u884c\u8f7b\u5ea6 CPU \u8fd0\u7b97\u4fdd\u6301 Render \u670d\u52a1\u6d3b\u8dc3\u3002", "operationId": "api_keepalive_render", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/keepalive-supabase": {"get": {"tags": ["Common"], "summary": "Api Keepalive Supabase", "description": "\u9632 Supabase \u4f11\u7720\u63a5\u53e3\u3002\n\n\u901a\u8fc7\u8f7b\u91cf SQL \u67e5\u8be2\u907f\u514d Supabase \u56e0\u957f\u671f\u65e0\u8bbf\u95ee\u8fdb\u5165\u4f11\u7720\u72b6\u6001\u3002", "operationId": "api_keepalive_supabase", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/keepalive-modal": {"get": {"tags": ["Common"], "summary": "Api Keepalive Modal", "description": "\u5b9a\u65f6\u89e6\u53d1\uff0c\u7528\u4e8e\u4fdd\u6301 Modal \u5bb9\u5668\u5b58\u6d3b\u72b6\u6001\uff0c\u9632\u6b62\u8d85\u65f6\u56de\u6536\u3002", "operationId": "api_keepalive_modal", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/proxy-predict": {"get": {"tags": ["Predict"], "summary": "Api Proxy Predict", "description": "\u4ee3\u7406\u63a8\u7406\u8bf7\u6c42\u63a5\u53e3\u3002\n\n\u5c06\u5ba2\u6237\u7aef\u8bf7\u6c42\u8f6c\u53d1\u81f3 Modal/GPU \u6a21\u578b\u670d\u52a1\uff0c\u652f\u6301 Token \u6821\u9a8c\u3002", "operationId": "api_proxy_predict", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PredictResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/template-meta": {"get": {"tags": ["Resource"], "summary": "Api Template Meta", "description": "\u6a21\u677f\u5143\u4fe1\u606f\u63a5\u53e3\u3002\n\n\u8fd4\u56de\u6240\u6709\u6a21\u677f\u7684\u7248\u672c\u53f7\u3001\u540d\u79f0\u4e0e\u4e0b\u8f7d\u5730\u5740\u3002", "operationId": "api_template_meta", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/toolkit-meta": {"get": {"tags": ["Resource"], "summary": "Api Toolkit Meta", "description": "\u5de5\u5177\u5143\u4fe1\u606f\u63a5\u53e3\u3002\n\n\u8fd4\u56de\u6240\u6709\u5de5\u5177\u7684\u7248\u672c\u53f7\u3001\u540d\u79f0\u4e0e\u4e0b\u8f7d\u5730\u5740\u3002", "operationId": "api_toolkit_meta", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}, {"name": "platform", "in": "query", "required": true, "schema": {"type": "string", "title": "Platform"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/model-meta": {"get": {"tags": ["Resource"], "summary": "Api Model Meta", "description": "\u6a21\u578b\u5143\u4fe1\u606f\u63a5\u53e3\u3002\n\n\u8fd4\u56de\u6240\u6709\u6a21\u578b\u7684\u7248\u672c\u53f7\u3001\u540d\u79f0\u4e0e\u4e0b\u8f7d\u5730\u5740\u3002", "operationId": "api_model_meta", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/template-viewer": {"get": {"tags": ["Resource"], "summary": "Api Template Viewer", "description": "\u5355\u4e2a\u6a21\u677f\u5185\u5bb9\u4e0b\u8f7d\u63a5\u53e3\u3002\n\n\u901a\u8fc7\u6a21\u677f\u540d\u83b7\u53d6\u5176\u7eaf\u6587\u672c\u5185\u5bb9\uff08\u5982 HTML\u3001JSON \u6a21\u677f\u7b49\uff09\u3002", "operationId": "api_template_viewer", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}, {"name": "page", "in": "query", "required": true, "schema": {"type": "string", "title": "Page"}}], "responses": {"200": {"description": "Successful Response", "content": {"text/plain": {"schema": {"type": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/business-case": {"get": {"tags": ["Resource"], "summary": "Api Business Case", "description": "\u83b7\u53d6\u4e1a\u52a1\u7528\u4f8b\u6307\u4ee4\u96c6\u3002\n\n\u6839\u636e `case` \u53c2\u6570\u8fd4\u56de\u4e00\u7ec4\u547d\u4ee4\uff0c\u7528\u4e8e\u5ba2\u6237\u7aef\u6267\u884c\u6d41\u7a0b\u914d\u7f6e\u3002", "operationId": "api_business_case", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}, {"name": "case", "in": "query", "required": true, "schema": {"type": "string", "title": "Case"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/self-heal": {"post": {"tags": ["SelfHeal"], "summary": "Api Self Heal", "description": "UI \u5143\u7d20\u81ea\u6108\u63a5\u53e3\uff0c\u4e00\u6b21\u6027\u8fd4\u56de\u6700\u7ec8\u7ed3\u679c\n\n\u57fa\u4e8e\u8bed\u4e49\u76f8\u4f3c\u5ea6\u81ea\u52a8\u5bfb\u627e\u6700\u53ef\u80fd\u7684\u65b0\u63a7\u4ef6\uff0c\u5b9e\u73b0\u5b9a\u4f4d\u4fee\u590d\u3002", "operationId": "api_self_heal", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/self-heal-stream": {"post": {"tags": ["SelfHeal"], "summary": "Api Self Heal Stream", "description": "UI \u5143\u7d20\u81ea\u6108 \u2014 \u6d41\u5f0f\u8fd4\u56de\u6267\u884c\u65e5\u5fd7\u4e0e\u6700\u7ec8\u51b3\u7b56\n\nStreamingResponse \u6587\u672c\u6d41\u8f93\u51fa:\n    [1] \u89e3\u6790\u8282\u70b9 \u2026\n    [2] \u751f\u6210\u5411\u91cf \u2026\n    [3] \u53ec\u56de \u2026\n    [4] \u91cd\u6392 \u2026\n    [\u5b8c\u6210] \u8fd4\u56de JSON \u7ed3\u679c", "operationId": "api_self_heal_stream", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/sign": {"post": {"tags": ["Signature"], "summary": "Api Sign", "description": "\u6388\u6743\u7b7e\u540d\u63a5\u53e3\u3002\n\n\u6839\u636e\u8bf7\u6c42\u4fe1\u606f\u751f\u6210\u7b7e\u540d\u8bc1\u4e66\uff08License \u6587\u4ef6\uff09\uff0c\u652f\u6301\u5ba2\u6237\u7aef\u6fc0\u6d3b\u9a8c\u8bc1\u3002", "operationId": "api_sign", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LicenseRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LicenseResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/speech-meta": {"get": {"tags": ["Speech"], "summary": "Api Speech Meta", "description": "\u83b7\u53d6\u8bed\u97f3\u5408\u6210\u683c\u5f0f\u5217\u8868\u3002\n\n\u8fd4\u56de\u53ef\u7528\u7684\u8bed\u97f3\u683c\u5f0f\u3001\u8bed\u8c03\u6a21\u578b\u4e0e\u8bed\u8a00\u8bbe\u7f6e\u3002", "operationId": "api_speech_meta", "parameters": [{"name": "a", "in": "query", "required": true, "schema": {"type": "string", "title": "A"}}, {"name": "t", "in": "query", "required": true, "schema": {"type": "integer", "title": "T"}}, {"name": "n", "in": "query", "required": true, "schema": {"type": "string", "title": "N"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/speech-voice": {"post": {"tags": ["Speech"], "summary": "Api Speech Voice", "description": "\u5408\u6210\u8bed\u97f3\u97f3\u9891\u6587\u4ef6\u3002\n\n\u63d0\u4ea4\u8bed\u97f3\u5185\u5bb9\u4e0e\u76ee\u6807\u683c\u5f0f\uff0c\u8fd4\u56de\u53ef\u4e0b\u8f7d\u7684\u97f3\u9891\u6587\u4ef6\u6216\u94fe\u63a5\u3002", "operationId": "api_speech_voice", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SpeechRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SpeechResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealRequest": {"properties": {"app_id": {"type": "string", "title": "App Id", "description": "\u5e94\u7528ID\uff0c\u5982\u5305\u540d/\u57df\u540d"}, "page_id": {"type": "string", "title": "Page Id", "description": "\u9875\u9762ID\uff0c\u4f8b\u5982 Activity/URL/\u8def\u7531"}, "platform": {"type": "string", "title": "Platform", "description": "android/web/ios"}, "old_locator": {"$ref": "#/components/schemas/Locator", "description": "\u539f\u59cb\u5b9a\u4f4d\u5bf9\u8c61"}, "page_dump": {"type": "string", "title": "Page Dump", "description": "\u9875\u9762dump\uff1aAndroid XML / Web HTML"}, "screenshot": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Screenshot", "description": "base64\u622a\u56fe(\u53ef\u9009)"}, "context": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Context", "description": "\u4e0a\u4e0b\u6587\u4fe1\u606f\uff0c\u4f8b\u5982intent/\u6d4b\u8bd5ID"}}, "type": "object", "required": ["app_id", "page_id", "platform", "old_locator", "page_dump"], "title": "HealRequest", "description": "Healing Request\n\nParameters\n----------\napp_id : str\n    \u5e94\u7528ID\uff0c\u4f8b\u5982\u5305\u540d/\u57df\u540d\npage_id : str\n    \u9875\u9762\u552f\u4e00ID\uff0c\u6bd4\u5982 Activity/URL/\u8def\u7531\nplatform : str\n    \u8fd0\u884c\u5e73\u53f0\uff0c\u53ef\u4e3a `android` `ios` `web`\nold_locator : Locator\n    \u539f\u59cb\u5b9a\u4f4d\u4fe1\u606f\u5bf9\u8c61\npage_dump : str\n    \u9875\u9762\u7ed3\u6784\u4fe1\u606f\uff0cAndroid\u4e3aXML\uff0cWeb\u4e3aHTML\nscreenshot : str, optional\n    \u9875\u9762\u622a\u56fe(Base64 PNG\uff0c\u53ef\u9009)\ncontext : dict, optional\n    \u4e0a\u4e0b\u6587\uff0c\u4f8b\u5982\u6d4b\u8bd5\u7528\u4f8bID\u3001intent\u3001\u81ea\u5b9a\u4e49\u53c2\u6570"}, "HealResponse": {"properties": {"healed": {"type": "boolean", "title": "Healed", "description": "\u662f\u5426\u6210\u529f\u81ea\u6108"}, "confidence": {"type": "number", "title": "Confidence", "description": "\u7f6e\u4fe1\u5ea6 0~1"}, "new_locator": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "New Locator", "description": "\u65b0\u5b9a\u4f4d\u5bf9\u8c61\uff0c\u4f8b\u5982 {'by':'bounds','value':'[100,200][300,400]'}"}, "details": {"additionalProperties": true, "type": "object", "title": "Details", "description": "\u8c03\u8bd5\u4fe1\u606f/\u5019\u9009\u5217\u8868\u7b49"}}, "type": "object", "required": ["healed", "confidence"], "title": "HealResponse", "description": "Healing Response\n\nReturns\n-------\nhealed : bool\n    \u662f\u5426\u6210\u529f\u4fee\u590d\u5b9a\u4f4d\nconfidence : float\n    \u7f6e\u4fe1\u5ea6 [0,1]\nnew_locator : dict, optional\n    \u4fee\u590d\u540e\u5b9a\u4f4d\u4fe1\u606f\uff0c\u5982 {\"by\":\"bounds\",\"value\":\"[100,200][300,400]\"}\ndetails : dict\n    \u8c03\u8bd5\u4fe1\u606f\u3001\u5019\u9009\u5b9a\u4f4d\u3001\u8bc4\u5206\u7b49"}, "LicenseRequest": {"properties": {"a": {"type": "string", "title": "A", "description": "\u5e94\u7528\u6807\u8bc6\uff0c\u4f8b\u5982\u5e94\u7528\u540d\u79f0\u6216\u5ba2\u6237\u7aefID"}, "t": {"type": "integer", "title": "T", "description": "\u8bf7\u6c42\u65f6\u95f4\u6233(Unix)\uff0c\u7528\u4e8e\u7b7e\u540d\u9a8c\u6743"}, "n": {"type": "string", "title": "N", "description": "\u8282\u70b9Device/Client\u6807\u8bc6\uff0c\u7528\u4e8e\u533a\u5206\u8bf7\u6c42\u53d1\u8d77\u8005"}, "code": {"type": "string", "title": "Code", "description": "\u6388\u6743\u7801\uff0c\u7528\u4e8e\u751f\u6210license\u7684\u79d8\u94a5\u51ed\u8bc1"}, "castle": {"type": "string", "title": "Castle", "description": "\u73af\u5883\u6807\u8bc6\uff0c\u4f8b\u5982 prod/test/dev"}, "license_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "License Id", "description": "\u7eed\u671f\u65f6\u4f20\u5165\u65e7LicenseID\uff0c\u65b0\u751f\u6210\u53ef\u4e0d\u4f20"}}, "type": "object", "required": ["a", "t", "n", "code", "castle"], "title": "LicenseRequest", "description": "License Creation Request\n\nParameters\n----------\na : str\n    \u5e94\u7528\u6807\u8bc6\uff0c\u53ef\u7528\u4e8e\u533a\u5206\u8c03\u7528\u6765\u6e90\u3002\nt : int\n    \u8bf7\u6c42\u65f6\u95f4\u6233\uff08Unix \u79d2\uff09\uff0c\u7528\u4e8e\u4e0e\u7b7e\u540d\u6821\u9a8c\u5173\u8054\u3002\nn : str\n    \u8bbe\u5907\u6216\u8282\u70b9\u552f\u4e00\u540d\u79f0\uff08client_id / node_name\uff09\u3002\ncode : str\n    \u6388\u6743\u7801\u6216\u51ed\u8bc1\u9a8c\u8bc1\u7801\uff0c\u7528\u4e8e\u540e\u53f0\u6821\u9a8c\u5408\u6cd5\u6027\u3002\ncastle : str\n    \u73af\u5883\u6807\u8bc6\uff0c\u4f8b\u5982 \"prod\" / \"test\" / \"dev\"\u3002\nlicense_id : str, optional\n    \u7eed\u671f\u65f6\u53ef\u643a\u5e26\u7684\u5df2\u6709 license ID\uff0c\u6ca1\u6709\u5219\u89c6\u4e3a\u65b0\u7533\u8bf7\u3002"}, "LicenseResponse": {"properties": {"data": {"type": "string", "title": "Data", "description": "Base64\u7f16\u7801\u540e\u7684 License \u6570\u636e\u8f7d\u8377(JSON encoded)"}, "signature": {"type": "string", "title": "Signature", "description": "Base64\u7f16\u7801\u7684\u6570\u5b57\u7b7e\u540d\uff0c\u7528\u4e8e\u6821\u9a8c\u6388\u6743\u5408\u6cd5\u6027"}}, "type": "object", "required": ["data", "signature"], "title": "LicenseResponse", "description": "License Response\n\nParameters\n----------\ndata : str\n    Base64\u7f16\u7801\u540e\u7684\u6388\u6743\u8f7d\u8377(JSON)\uff0c\u9700\u8981\u5ba2\u6237\u7aef\u89e3\u7801\u4ee5\u83b7\u53d6\u5b8c\u6574\u5185\u5bb9\u3002\nsignature : str\n    Base64\u7f16\u7801\u7684\u7b7e\u540d\uff0c\u7528\u4e8e\u9a8c\u8bc1license\u5b8c\u6574\u6027\u4e0e\u5408\u6cd5\u6765\u6e90\u3002"}, "Locator": {"properties": {"by": {"type": "string", "title": "By", "description": "\u5b9a\u4f4d\u65b9\u5f0f\uff0c\u5982 id/xpath/css/bounds"}, "value": {"type": "string", "title": "Value", "description": "\u5b9a\u4f4d\u503c"}}, "type": "object", "required": ["by", "value"], "title": "Locator", "description": "Locator\n\nParameters\n----------\nby : str\n    \u5b9a\u4f4d\u65b9\u5f0f\uff0c\u5982 `id` / `xpath` / `css` / `bounds`\nvalue : str\n    \u4e0e\u5b9a\u4f4d\u65b9\u5f0f\u5339\u914d\u7684\u503c"}, "PredictResponse": {"properties": {"configuration": {"additionalProperties": true, "type": "object", "title": "Configuration", "description": "\u63a8\u7406\u914d\u7f6e\u5bf9\u8c61", "default": {}}, "available": {"type": "boolean", "title": "Available", "description": "\u63a8\u7406\u670d\u52a1\u662f\u5426\u53ef\u7528"}, "expire_at": {"type": "integer", "title": "Expire At", "description": "\u8fc7\u671f\u65f6\u95f4\u6233"}, "timeout": {"type": "number", "title": "Timeout", "description": "\u63a8\u7406\u8bf7\u6c42\u8d85\u65f6(\u79d2)", "default": 60.0}, "content_type": {"type": "string", "title": "Content Type", "description": "\u8bf7\u6c42\u5185\u5bb9\u7c7b\u578b", "default": "multipart/form-data"}, "auth_header": {"type": "string", "title": "Auth Header", "description": "\u9274\u6743\u5934\u683c\u5f0f"}, "token": {"type": "string", "title": "Token", "description": "\u7b7e\u540d Token"}, "method": {"type": "string", "title": "Method", "description": "\u8bf7\u6c42\u65b9\u5f0f", "default": "POST"}, "url": {"type": "string", "title": "Url", "description": "\u63a8\u7406\u8c03\u7528\u5730\u5740"}, "ttl": {"type": "integer", "title": "Ttl", "description": "\u5269\u4f59\u6709\u6548\u671f(\u79d2)"}, "region": {"type": "string", "title": "Region", "description": "\u533a\u57df\u6807\u8bc6"}, "version": {"type": "string", "title": "Version", "description": "\u7248\u672c\u53f7"}, "message": {"type": "string", "title": "Message", "description": "\u54cd\u5e94\u72b6\u6001\u6d88\u606f", "default": "Predict service online"}}, "type": "object", "required": ["available", "expire_at", "auth_header", "token", "url", "ttl", "region", "version"], "title": "PredictResponse", "description": "Predict Service License Response\n\nParameters\n----------\nconfiguration : dict\n    \u63a8\u7406\u670d\u52a1\u914d\u7f6e\u5185\u5bb9\uff08\u9884\u7559\u5b57\u6bb5\uff0c\u9ed8\u8ba4\u4e3a\u7a7a\u5bf9\u8c61\uff09\u3002\navailable : bool\n    \u63a8\u7406\u670d\u52a1\u662f\u5426\u53ef\u7528\u3002\nexpire_at : int\n    license \u5230\u671f\u65f6\u95f4\u6233 (Unix Time)\u3002\ntimeout : float\n    \u8bf7\u6c42\u6700\u5927\u7b49\u5f85\u65f6\u95f4\uff0c\u5355\u4f4d\u79d2\u3002\ncontent_type : str\n    \u63a8\u7406 API \u671f\u671b\u7684 Content-Type\u3002\nauth_header : str\n    \u9274\u6743\u5934\u683c\u5f0f\uff0c\u4f8b\u5982 \"Bearer {token}\"\u3002\ntoken : str\n    \u751f\u6210\u7684\u8bbf\u95ee\u7b7e\u540d Token\u3002\nmethod : str\n    \u63a8\u7406\u8c03\u7528\u4f7f\u7528\u7684\u8bf7\u6c42\u65b9\u6cd5\uff08\u901a\u5e38 POST\uff09\u3002\nurl : str\n    \u63a8\u7406\u670d\u52a1\u8c03\u7528\u5730\u5740\u3002\nttl : int\n    \u5f53\u524d license \u6709\u6548\u671f\u79d2\u6570\u3002\nregion : str\n    \u670d\u52a1\u533a\u57df\u6807\u8bc6\uff0c\u4f8b\u5982 \"us-east-1\"\u3002\nversion : str\n    \u63a8\u7406\u670d\u52a1\u7248\u672c\u53f7\u3002\nmessage : str\n    \u8fd4\u56de\u72b6\u6001\u4fe1\u606f\u3002"}, "SpeechRequest": {"properties": {"a": {"type": "string", "title": "A", "description": "\u5e94\u7528ID\u7b7e\u540d\u5b57\u6bb5"}, "t": {"type": "integer", "title": "T", "description": "\u65f6\u95f4\u6233(\u79d2\u7ea7)\uff0c\u7528\u4e8e\u8bf7\u6c42\u6821\u9a8c"}, "n": {"type": "string", "title": "N", "description": "\u968f\u673anonce\uff0c\u9632\u91cd\u590d\u8bf7\u6c42"}, "speak": {"type": "string", "title": "Speak", "description": "\u6587\u672c\u5185\u5bb9\uff0c\u5c06\u88ab\u5408\u6210\u4e3a\u97f3\u9891"}, "voice": {"type": "string", "title": "Voice", "description": "\u8bed\u97f3\u540d\u79f0", "default": "zh-CN-XiaoxiaoNeural"}, "waver": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Waver", "description": "\u8f93\u51fa\u97f3\u9891\u683c\u5f0f\uff0c\u5982 mp3/wav", "default": "mp3"}, "rater": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Rater", "description": "\u8bed\u901f\uff0c\u5982 +20%\u3001-10%", "default": "0%"}, "pitch": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pitch", "description": "\u8bed\u8c03\uff0c\u5982 +5%", "default": "0%"}, "volume": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Volume", "description": "\u97f3\u91cf\uff0c\u5982 +0dB", "default": "default"}, "manner": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Manner", "description": "\u60c5\u611f\u98ce\u683c\uff0c\u5982 cheerful"}, "degree": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Degree", "description": "\u98ce\u683c\u5f3a\u5ea6\uff0c\u5982 1.0\u30012.0"}}, "type": "object", "required": ["a", "t", "n", "speak"], "title": "SpeechRequest", "description": "Speech Request\n\nParameters\n----------\na : str\n    \u5e94\u7528ID\u7b7e\u540d\u5b57\u6bb5\uff08\u53c2\u4e0e\u9274\u6743\uff09\nt : int\n    \u65f6\u95f4\u6233\uff08\u79d2\u7ea7\uff09\nn : str\n    \u968f\u673anonce\uff0c\u9632\u91cd\u653e\nspeak : str\n    \u6587\u672c\u5185\u5bb9\uff0c\u5c06\u5408\u6210\u4e3a\u8bed\u97f3\nvoice : str, default: \"zh-CN-XiaoxiaoNeural\"\n    \u4f7f\u7528\u7684\u8bed\u97f3\u540d\u79f0\nwaver : str, optional, default: \"mp3\"\n    \u8f93\u51fa\u97f3\u9891\u683c\u5f0f\uff0c\u5982 `mp3` / `wav`\nrater : str, optional, default: \"0%\"\n    \u8bed\u901f\u8bbe\u7f6e\uff0c\u4f8b\u5982 `+20%`\u3001`-10%`\npitch : str, optional, default: \"0%\"\n    \u97f3\u8c03\u8bbe\u7f6e\uff0c\u5982 `+5%`\nvolume : str, optional, default: \"default\"\n    \u97f3\u91cf\u8bbe\u7f6e\uff0c\u4f8b `+0dB`\nmanner : str, optional\n    \u60c5\u7eea\u8868\u8fbe\u65b9\u5f0f\uff0c\u5982 `cheerful`\ndegree : str, optional\n    \u98ce\u683c\u5f3a\u5ea6\uff0c\u5982 `1.0`, `2.0`"}, "SpeechResponse": {"properties": {"url": {"type": "string", "title": "Url", "description": "\u8bed\u97f3\u97f3\u9891\u6587\u4ef6\u7b7e\u540d\u4e0b\u8f7d\u5730\u5740\uff0c\u7528\u4e8e\u76f4\u63a5\u8bbf\u95ee\u97f3\u9891\u8d44\u6e90", "examples": ["https://cdn.xxx.com/audio/20250101/voice_xxx.wav?token=abc123"]}}, "type": "object", "required": ["url"], "title": "SpeechResponse", "description": "Response Model (Return Speech Audio URL)\n\nParameters\n----------\nurl : str\n    \u8bed\u97f3\u6587\u4ef6\u7684\u7b7e\u540d\u4e0b\u8f7d\u5730\u5740\uff0c\u7528\u4e8e\u5ba2\u6237\u7aef\u76f4\u63a5\u8bbf\u95ee\u3002"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}